Workaround occasional crashes using canary (temporary fix)
When running in a real-life 802.11a Wi-Fi environment with busy traffic,
the module occasionally crashes (segfault) and/or stops responding.
This was traced back to a out-of-bounds memory write in class
decode_mac_impl. I don't know the root cause.

As a temporary workaround, I moved around some class members that were
getting overwritten and added a canary to demonstrate the out-of-bounds
memory write. The module runs stable and doesn't crash after these
modifications.

Signed-off-by: Jonathan Brucker <jonathan.brucke@gmail.com>